re-organize providers/component modules
