+++ cmd/exo/new_process.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/deref/exo/components/process"
 	"github.com/deref/exo/core/api"
+	"github.com/deref/exo/providers/unix/components/process"
 	"github.com/deref/exo/util/jsonutil"
 	"github.com/spf13/cobra"
 )
