+++ cmd/exop/main.go
@@ -8,11 +8,11 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/deref/exo/components/log"
 	"github.com/deref/exo/core/server"
 	"github.com/deref/exo/core/state/statefile"
 	josh "github.com/deref/exo/josh/client"
 	logd "github.com/deref/exo/logd/client"
+	"github.com/deref/exo/providers/core/components/log"
 	"github.com/deref/exo/supervise"
 	"github.com/deref/exo/util/cmdutil"
 	"github.com/deref/exo/util/httputil"
