+++ core/server/workspace.go
@@ -8,18 +8,18 @@ import (
 	"strings"
 
 	"github.com/deref/exo/chrono"
-	"github.com/deref/exo/components/docker/container"
-	"github.com/deref/exo/components/docker/network"
-	"github.com/deref/exo/components/docker/volume"
-	"github.com/deref/exo/components/invalid"
-	"github.com/deref/exo/components/log"
-	"github.com/deref/exo/components/process"
 	"github.com/deref/exo/core/api"
 	core "github.com/deref/exo/core/api"
 	state "github.com/deref/exo/core/state/api"
 	"github.com/deref/exo/gensym"
 	logd "github.com/deref/exo/logd/api"
 	"github.com/deref/exo/manifest"
+	"github.com/deref/exo/providers/core/components/invalid"
+	"github.com/deref/exo/providers/core/components/log"
+	"github.com/deref/exo/providers/docker/components/container"
+	"github.com/deref/exo/providers/docker/components/network"
+	"github.com/deref/exo/providers/docker/components/volume"
+	"github.com/deref/exo/providers/unix/components/process"
 	"github.com/deref/exo/util/errutil"
 	"github.com/deref/exo/util/jsonutil"
 	docker "github.com/docker/docker/client"
